Finish 06-flex-layout
Took a little while to do this one - great chance to experiment with
what putting things in different places does to the end result. Did a
few things slightly differently but nothing that altered the outcome.
Also put in a few pieces of redundant code which is worth being aware
of to prevent adding things in unnecessarily.